+++ vcfpy/__init__.py
@@ -13,8 +13,8 @@ from .header import Header, HeaderLine, SimpleHeaderLine, \
     PedigreeHeaderLine, SampleHeaderLine, CompoundHeaderLine, InfoHeaderLine, \
     FormatHeaderLine, SamplesInfos, FieldInfo, header_without_lines
 
-from .record import Record, Call, AltRecord, Substitution, BreakEnd, \
-    SingleBreakEnd, SymbolicAllele
+from .record import Record, Call, UnparsedCall, AltRecord, Substitution, \
+    BreakEnd, SingleBreakEnd, SymbolicAllele
 
 from .record import SNV, MNV, DEL, INS, INDEL, SV, BND, SYMBOLIC, MIXED
 from .record import HOM_REF, HET, HOM_ALT